feat: Added row selection feature to DataTable component
Updated the DataTable component to include a row selection feature. Also included changes in the AppRouter.js to have create and view all Technologies routed correctly. A new file all.js is added to display all the technologies. Minor code formatting and removal of extra white spaces are also part of this commit.