Refactor campaign events view and create technology info
This commit refactors the campaign events view and create technology info components. It updates the handleRemove function in the CampaignEventsView component to correctly remove events and update the campaign technologies. It also updates the updateTechObject function in the Info component to pass the correct argument. Additionally, it adds a call to onCancel in the Testimonials component after creating a testimonial.