fix: Static texts translation (#385)
* Add notification messages and headers to multiple languages

Included success and error notifications for testimonials and comments in English, Spanish, Portuguese, and French. Added new headers for sections such as "Newsletter," "Getting Started," and "Meet the Coaches" across all supported languages.

* Update coach section title to include static text

Modified the `CoachesSectionWithFilters` component to display the `staticT?.title` in the section title if customization is not available. This ensures a default title is shown even if customization data is missing.

* Use static text variables for user notifications

Replaced hardcoded text strings with static text variables in both CommentComponentForModal.js and Footer.js to improve maintainability and support for internationalization. This ensures that all user-facing text can be easily managed and translated as needed.

* Enhance localization support with static text integration

Improved localization by integrating static text placeholders across multiple pages. This includes adjustments to trigger functions, form notifications, and section titles to utilize static text sources when available.

* Update user testimonial formatting

Modified the format for displaying the user's company name in testimonials to include a static translation key. This ensures the word 'from' can be localized correctly based on user settings.

* Remove TODO comment in Footer.js

Eliminated an outdated TODO comment related to translation help text in the footer component. This will improve code readability and reduce unnecessary clutter.

* Update text property access in components

Modified `CoachesSectionWithFilters.js` and `Footer.js` to access the `text` property of nested objects in the `title` and `description` fields. This ensures the correct text is displayed based on the provided customization or static data.

* Refactor JSON statics structure to improve consistency

Updated the internationalization JSON structure to nest titles and descriptions under 'text' keys. This change ensures a consistent format across multiple sections and languages, enhancing maintainability and readability of the code. Adjusted relevant component render logic to accommodate this new structure.